refactor: moved calender and related component to students page
