--- src/pages/attendence/attendence.tsx
@@ -1,8 +1,0 @@
-import AttendenceList from "@/components/attendence/attendence";
-
-// Attendence list page Component
-function AttendenceListPage() {
-    return <AttendenceList />;
-}
-
-export default AttendenceListPage;
