+++ src/pages/student/attendence.tsx
@@ -1,0 +1,8 @@
+import AttendenceList from "@/components/student/attendence/attendence";
+
+// Attendence Page Component
+function AttedencePage() {
+    return <AttendenceList />;
+}
+
+export default AttedencePage;
