--- src/pages/student/leave.tsx
@@ -1,8 +1,0 @@
-function Leave() {
-    return (
-        <div>
-        </div>
-    );
-}
-
-export default Leave;
